Moved 10-deploy to 10-deploy.py
The default behavior of the sentry.wait() method was causing some false
positives in CWR. I've switched over to wait_for_messages() to sniff out
the ETCD deployment status and only proceed once Healthy/Unhealthy has
been reported.

One might argue why I didn't only decorate for healthy. I feel that the
output of the tests is still going to be helpful in determining the
severity of the failure, and might give insight into why the deployment
is unhealthy.