Adds the snapshot action
Half of the puzzle of migrating etcd. The snapshot action allows a
cluster admin to export a point-in-time snapshot of the etcd data
residing in the cluster.

Only a single node will need to export its data, and only a single node
can re-import and start a new cluster. For more information about this
process see: https://coreos.com/etcd/docs/latest/admin_guide.html