Adds the restore action
This adds the capability of the etcd charm to restore the state of an
etcd cluster created from the `snapshot` action. The process is
documented on the etcd admin guide:
https://coreos.com/etcd/docs/latest/admin_guide.html

Note: that restoring from a snapshot is only supported on new clusters,
as restoring from a snapshot on a scaled cluster can yield a broken
installation. Peers are dropped and the cluster ID is changed during new
cluster initialization.