Attempted fix for #70
This allows the member listing during certificate exchange to fail, due
in part to having no control over the masters state, which was causing
transient failures during turnup while the master etcd service was
recycling during initial deployment.

There are likely additional code paths that will need this style of
guard, but this was the lowest impact fix vector for this issue.