#85: Resolve issues with cluster scale when removing the leader (fixes kubernetes/kubernetes#43461)
This branch removes the erronious logic when tracking the -departed
relation states to remove all participating peers. Instead we are now
relying on the behavior of the -broken hook to only execute on the unit
that is being removed from the cluster. This may not be perfect but
works 100% better than the current implementation.