support new snapd_refresh from layer-snap
Default our etcd charm to use the 'max' possible refresh window,
which is currently to refresh once a month. The leader will get
the initial value and set leader data so all followers (present
and future) will refresh around the same time.

Adding an amulet test for this and fixed up lint issues in the tests.