gate snapd_refresh handlers on snap.refresh.set
If snapd_refresh handlers from layer-etcd run before layer-snap,
followers may set the refresh config to the leader's data and
subsequently have it overwritten by layer-snap.

Prevent this by gating our refresh handlers with 'snap.refresh.set':

https://git.launchpad.net/layer-snap/commit/?h=defer-refresh&id=62d918f8979451128af93973aa57635e19287fee

This ensures layer-snap is done configuring the core snap before we
process our leader config. This is also useful on the leader so we
don't bother trying to get the config until we know it's set.