support new snapd_refresh from layer-snap (#124)
* support new snapd_refresh from layer-snap

Default our etcd charm to use the 'max' possible refresh window,
which is currently to refresh once a month. The leader will get
the initial value and set leader data so all followers (present
and future) will refresh around the same time.

Adding an amulet test for this and fixed up lint issues in the tests.

* snap.get works now, use it

* gate snapd_refresh handlers on snap.refresh.set

If snapd_refresh handlers from layer-etcd run before layer-snap,
followers may set the refresh config to the leader's data and
subsequently have it overwritten by layer-snap.

Prevent this by gating our refresh handlers with 'snap.refresh.set':

https://git.launchpad.net/layer-snap/commit/?h=defer-refresh&id=62d918f8979451128af93973aa57635e19287fee

This ensures layer-snap is done configuring the core snap before we
process our leader config. This is also useful on the leader so we
don't bother trying to get the config until we know it's set.