Fix health action
Health action is currently implemented using a bash script that
queries etcdv2's cluster-health, but it is currently failing
because the API version is not being defined by the script, so
on newer versions of etcd it fails since it defaults to v3.

This patch fixes it by changing the action to use the EtcdCtl
wrapper class.

Fixes: https://bugs.launchpad.net/charm-etcd/+bug/1911437