Remove ubuntu user access to etcd tls cert files
This is for general security hardening.
It means the ubuntu user will no longer have access
to the sensitive tls cert, key, and ca files for etcd
(ie. the files in /var/snap/etcd/common/ ).

An implication of this is that the `etcdctl` command
will no longer be configured or usable for the ubuntu user.
In cases where users previously accessed the shell
on an etcd unit to run etcdctl commands as the ubuntu user,
they will now need to run as root from an interactive shell
(interactive because the config is set in .bash_aliases).

This also means that for any cases where users manually copied
(via ssh or scripts) the cert files from an etcd unit,
they will now need to use sudo.

Interestingly, etcdctl from the local etcd unit with tls
has been broken anyway,
due to 127.0.0.1 not being included in the etcd certs.
See https://github.com/charmed-kubernetes/layer-etcd/pull/204

Closes-Bug: https://bugs.launchpad.net/charm-etcd/+bug/1997531